requirements.txt
streamlit
google-generativeai
python-dotenv
selenium
beautifulsoup4